Fixed Commit::getLastModification
If we use gitlib (located in /a/b/gitlib) and try to inspect a
project located in /a/b ; Gitlib will not be able to get the log
since it run a command like:

    git log --working-dir /a/b/.git --working-tree /a/b log --format=%H -n 1 THEHASH -- PATH

so PATH does not exist in /a/b/gitlib/ (ie /a/b/gitlib/PATH).
so the command fails.

So i fixed it by preprend the working-dir.

Note: I think we should update all commands that use a path in arguments.

ping @alexandresalome